refactors to move gemini to seperate module
