+++ src/index.ts
@@ -3,17 +3,8 @@ import { promises as fs } from 'fs';
 import { exec } from 'child_process';
 import path from 'path';
 import os from 'os';
-import { GoogleGenerativeAI } from '@google/generative-ai';
-
-type Config = {
-  apiKey: string;
-  model: string;
-};
-
-interface ShellInfo {
-  executable: string;
-  friendlyName: string;
-}
+import { ShellInfo, Config } from './types';
+import * as gemini from './gemini';
 
 const configDirectoryName = '.genshell';
 const configFileName = 'genshell_config.json';
@@ -97,47 +88,13 @@ async function generateBashCommand(description: string): Promise<string> {
   const cfg = await loadConfig();
   const shellInfo = getShellInfo();
   const osName = os.platform();
-  const geminiAPIKey = cfg.apiKey;
-  const genAI = new GoogleGenerativeAI(geminiAPIKey);
-  const model = genAI.getGenerativeModel({ model: cfg?.model ?? defaultModel });
-  const prompt = `
-  You are a smart bot that can generate ${shellInfo.friendlyName} commands on ${osName} from a description provided by user. \
-
-  ## Examples:
-  In this section, you find some examples, what user includes in the description and what the bot generates.
-
-  Example 1:
-  Description: "List all files in the current directory."
-  Generated command: "ls -al"
-
-  Example 2:
-  Description: "Create a new directory named 'my-dir'."
-  Generated command: "mkdir my-dir"
-
-  Example 3:
-  Description: "Print the current date."
-  Generated command: "date"
-
-  ## Rules
-  In this section you find rules that you need to follow:
-
-  Rule 1: The command should be a valid ${shellInfo.friendlyName} command that can be executed on ${osName}.
-  Rule 2: The command should be relevant to the description provided by the user.
-  Rule 3: Your output should include the command generated by the bot, or some friendly message if no command is generated.
-  Rule 4: Always rely on official documentation for the commands you generate.
-
-  ## Input
-  Description: "${description}"
-  Shell: ${shellInfo.friendlyName}
-  OS: ${osName}
-
-  ## Output
-  executable shell command or friendly message, no extra message in the execution output.
-  `;
-
-  const result = await model.generateContent(prompt);
-  const response = await result.response;
-  const text = response.text();
+  const text = await gemini.generateBashCommand({
+    apiKey: cfg.apiKey,
+    shellInfo,
+    osName,
+    description,
+    model: cfg.model,
+  });
   if (typeof text !== 'string') {
     return 'No command generated.';
   }
