+++ src/utils.ts
@@ -1,0 +1,6 @@
+import os from 'os';
+import { OsName } from './types';
+
+export function getOsName(): OsName {
+  return os.platform();
+}
